Fix: code style and deliveTaget interface apply
