+++ src/interface/deliveryTarget.ts
@@ -15,35 +15,3 @@ export type DeliveryTarget = {
     vpcID?: string;
   };
 };
-
-export interface QueryDeliveryTarget {
-  query?: string;
-  page?: number;
-  pageSize?: number;
-}
-
-export interface DeliveryTargetState {
-  deliveryTagetList: [];
-  deliveryTagetListTotal: number;
-}
-
-export interface ModelsType {
-  namespace: string;
-  state: DeliveryTargetState;
-  effects: {};
-  reducers: {};
-}
-
-export interface ResponseGenerator {
-  config?: any;
-  data?: any;
-  headers?: any;
-  request?: any;
-  status?: number;
-  error?: any;
-  statusText?: string;
-  records?: any;
-  pageNum?: any;
-  deliveryTagetList?: any;
-  deliveryTagetListTotal?: any;
-}
