+++ src/model/deliveryTarget.ts
@@ -1,10 +1,36 @@
 import { getDeliveryTarget, createDeliveryTarget } from '../api/deliveryTarget';
-import {
-  ModelsType,
-  ResponseGenerator,
-  DeliveryTargetState,
-  QueryDeliveryTarget,
-} from '../interface/deliveryTarget';
+
+export interface QueryDeliveryTarget {
+  query?: string;
+  page?: number;
+  pageSize?: number;
+}
+
+interface DeliveryTargetState {
+  deliveryTagetList: [];
+  deliveryTagetListTotal: number;
+}
+
+interface ModelsType {
+  namespace: string;
+  state: DeliveryTargetState;
+  effects: {};
+  reducers: {};
+}
+
+interface ResponseGenerator {
+  config?: any;
+  data?: any;
+  headers?: any;
+  request?: any;
+  status?: number;
+  error?: any;
+  statusText?: string;
+  records?: any;
+  pageNum?: any;
+  deliveryTagetList?: any;
+  deliveryTagetListTotal?: any;
+}
 
 const DeliveryTargets: ModelsType = {
   namespace: 'deliveryTarget',
